refactor: moved saveAutomaton function into mutations.ts
